Update Solution to properly pack design tools
When the Controls project was exploded into several sub-projects,
they were not marked for build in the solution file, leading to the
missing of design tools assemblies in their respective control's packages.

This patch updates the solution file with build options similar to what
was already working, before the splitting of the Controls project.